Move calendar to common component
