+++ src/app/_components/common/calendar/calendar-view.module.scss
