+++ src/app/_components/common/calendar/calendar-view.tsx
@@ -7,14 +7,14 @@ import timeGridPlugin from "@fullcalendar/timegrid";
 import { Box, Text } from "@mantine/core";
 import { useMediaQuery } from "@mantine/hooks";
 import { useEffect, useMemo, useRef } from "react";
+import Check from "@/assets/icons/check";
+import Cross from "@/assets/icons/cross";
 import {
   TABLE_SLOT_DURATION,
   TABLE_SLOT_MAX_TIME,
   TABLE_SLOT_MIN_TIME,
 } from "@/constants/TableScheduleConstants";
-import Check from "../../../assets/icons/check";
-import Cross from "../../../assets/icons/cross";
-import { type Booking, BookingStatus, type CalendarEvent } from "../../../types/types";
+import { type Booking, BookingStatus, type CalendarEvent } from "@/types/types";
 import styles from "./calendar-view.module.scss";
 
 // Event color constants
